udacity first project
Project Overview
Are you starting to feel like a web developer yet? You should!

Before you move on, I have one more challenge. I want you to use what you've learned up to this point to complete the following project. Simply titled, "Animal Trading Cards", this project combines the skills you mastered in the problem set and asks you to re-create a webpage from a design prototype.

This is a common workflow for front-end web developers. Typically, you'll be provided with a design prototype that needs to be translated to an actual, functional website. In most cases, designers only provide you with the design prototype. However, for this project, I've provided you with the design prototype and the HTML. The design prototype used in this project is inspired by trading cards and features a fish you might recognize from a popular animated film. You’ll be creating the card and swapping out the fish with an animal of your choice.